location_modeling_ui/backend/requirements.txt
Flask==2.3.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
osmnx==1.6.0
networkx==3.1
pulp==2.7.0
geopandas==0.13.2
matplotlib==3.7.2
contextily==1.3.0
shapely==2.0.1
